refactor: swap (#3164)
* refactor: swap

* refactor: reduce files
