--- src/providers/okx/constant.ts
@@ -1,3 +1,0 @@
-export const OKX_HOST = 'https://okx-proxy.r2d2.to';
-
-export const NATIVE_TOKEN_ADDRESS = '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE';
